requirements.txt
flask
flask_cors
requests
click
Flask_SocketIO==4.3.1
python-engineio==3.13.2
python-socketio==4.6.0
redis
geopy==2.1.0
python-dotenv
